requirements.txt
flask
spotipy
yt-dlp
python-dotenv
mutagen
